KNXexIP.GenStageProducer (knxex_ip v0.1.0)

KNX telegrams GenStage producer and broadcast dispatcher.

Functions

start_link(opts \\ [])

@spec start_link(Keyword.t()) :: {:ok, pid()} | {:error, {:already_started, pid()}}

Starts the GenStage producer and broadcast dispatcher.

subscribe_to(knx_interface, name_or_pid \\ nil, opts \\ [])

@spec subscribe_to(module(), atom() | pid(), Keyword.t()) :: :ok

Subscribes to the given KNX interface and thus receives KNX telegrams from it.

The received KNX telegrams are forwarded as GenStage events, in the form of {:knx, KNXexIP.Telegram.t()}.

This producer can also subscribe to GroupAddressServer, which will give you better information about the group address, but less information about the source. However said server uses a different struct and does not emit telegrams. You will receive {:knx_group, GroupAddressData.t()} events in that case.

unsubscribe_from(knx_interface, name_or_pid \\ nil, opts \\ [])

@spec unsubscribe_from(module(), atom() | pid(), Keyword.t()) :: :ok

Unsubscribes from the given KNX interface. KNX telegrams cannot be received from it anymore.